LayoutTests/media/picture-in-picture/picture-in-picture-api-exit-pip-1-expected.txt

This tests that exit Picture-in-Picture resolves when there is a Picture-in-Picture video.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(internals.setMockVideoPresentationModeEnabled(true))
RUN(video.src = findMediaFile("video", "../content/test"))
EVENT(canplaythrough)
EVENT(enterpictureinpicture)
END OF TEST
